Add files via upload
This is code I wrote to demonstrate the functionality of 1/32 scale smart home that utilizes both an Arduino and a Raspberry Pi. The code demonstrates the functionality of the smart home by turning on lights, turning on a fan when the temperature becomes to high, and closing the blinds automatically when the sun sets.